# src/Domain/Obras/Enums.cs
namespace Plataforma.Domain.Obras;

public enum EstadoProyecto
{
    Planificacion,
    EnEjecucion,
    Pausado,
    Finalizado
}

public enum EstadoHito
{
    Pendiente,
    EnProgreso,
    Completado
}